Module elorapi::connections

This module is for the connection to the Chirpstack server and a MQTT broker.
It is necessary to establish these connections to get the full functionality of this library.

Structs

ChirpstackConnection

This is to manage the connection to the Chirpstack Application Server.

Mqtt

This is to manage the connection to a Mqtt broker.