Module elorapi::devices

This module is for the management of devices and device profiles.

Note that it is necessary to establish a connection before loading data from or writing to the Chirsptack server. Beware that DeviceProfileContainer::establish_connection establishes another connection than DeviceContainer::establish_connection, therefor these functions must be called in their respective cases.

Also note that there is a difference between a DeviceProfile and a DeviceProfileListItem, which is here referred to as chirpstack device profile.
The same applies to Devices and DeviceListItems, which are referred to as chirpstack devices.

In the following the usage of this module will be described with an example.

Example

Startup

Creating a device profile with a specification file

Loading existing device profile

Loading specification to an existing device profile

Loading a device

Startup

It is necessary to create a device profile container and a device container.
To read from or to write to the Chrisptack sever a ChirpstackConnection is needed. For more information see the connections module.

use std::sync::{Arc, Mutex};
use elorapi::connections::ChirpstackConnection;
use elorapi::devices::{DeviceContainer, DeviceProfileContainer};

// create a DeviceContainer
let dev_container_arc_mutex = Arc::new(Mutex::new(DeviceContainer::new()));
let dev_container_arc = Arc::clone(&dev_container_arc_mutex);
let mut dev_container = dev_container_arc.lock().unwrap();

// creat a DeviceProfileContainer
let dev_prof_container_arc_mutex = Arc::new(Mutex::new(DeviceProfileContainer::new()));
let dev_prof_container_arc = Arc::clone(&dev_prof_container_arc_mutex);
let mut dev_prof_container = dev_prof_container_arc.lock().unwrap();

// create a ChirpstackConnection
let connection = ChirpstackConnection::new("API_Token", "server_uri");

Creating a device profile with a specification file

To create a new device profile, it is possible to read all necessary specifications out of a json file. For that the network server id and the organization id, to which the profile should be written, is necessary. The ids can be seen in the Chirpstack application server.

For more information about the specification files go to the specification files paragraph.

use elorapi::devices::DeviceProfile;

// the network server id and the organization id
let network_server_id = 11;
let organization_id = 1;
// path of the specification file
let path = "./specification_file_path/specification_file.json";

// read device profile out of specification file
let mut device_profile = DeviceProfile::read_specification(path, network_server_id, organization_id).unwrap();

// write the device profile to the Chirpstack server
device_profile.write_device_profile(connection.clone()).await.unwrap();

// add device profile to the container
dev_prof_container.add_device_profile(device_profile);

Loading existing device profile

It is possible to load existing device profiles out of the Chirpstack server.
For this the following steps need to be done:

Establish a device profile container connection
Note that this is another connection than the Chirpstack connection.

dev_prof_container.establish_connection(connection.clone()).await.unwrap();

Load chirpstack device profile
Note that Chirpstack device profiles should not be misunderstood as DeviceProfile.
It is only possible to load a specific number of device profiles from a specific organization and application. Therefor the specific ids must be given (this information can be seen in the Chirpstack application).

// number of items and ids
let limit = 10;
let organization_id = 1;
let application_id = 1;

// load the device profiles and add them to the container
dev_prof_container.load_chirpstack_device_profiles(limit, organization_id, application_id, connection.clone()).await.unwrap();

Load actual device profile
After getting the chirpstack device profile the actual device profile can now be loaded. For this the device profile id is needed (to get this information, either print the item via DeviceProfileContainer::print_list_items or get all items via DeviceProfileContainer::get_chirpstack_device_profiles and then their id).

 use elorapi::devices::DeviceProfile;

 // get the device profile id of a chirpstack device list item
 let device_profile_id = dev_prof_container.get_chirpstack_device_list()[0].id;

 // load the actual device profile form chirpstack
 let new_device_profile = DeviceProfile::load_device_profile(device_profile_id, connection.clone()).await.unwrap();

 // add the device profile to the container
 dev_prof_container.add_device_profile(new_device_profile);

Loading specification to an existing device profile

It is possible, for some cases even necessary, to load uplink and/or downlink specification/s via a specification file. It can be done directly after loading the device profile.

 use elorapi::devices::DeviceProfile;

 // load device profile out from the server
 let mut new_device_profile = DeviceProfile::load_device_profile("device_profile_id", connection.clone()).await.unwrap();
 // read the specification file for downlink, uplink or both
 new_device_profile.read_downlink("./specification_file_path/downlink_specification.json").unwrap();
 new_device_profile.read_uplink("./specification_file_path/uplink_specification.json").unwrap();

 // add device profile to the container
 dev_prof_container.add_device_profile(new_device_profile);

It is also possible to load the specification after loading an existing device profile. For this the index of the device profile in the container is needed.

// get the index of the device profile
let index = dev_prof_container.get_device_profile_index_via_dev_prof_id("device_profile_id").unwrap();
// now read the downlink or uplink specification file
dev_prof_container.get_device_profiles()[index].read_uplink("./specification_file_path/uplink_specification_file.json").unwrap();

Loading a device

It is possible to load existing devices from the Chirpstack server.
For this the following steps need to be done:

Establish a device container connection

 // establish a connection to the Chripstack server
 dev_container.establish_connection(connection.clone()).await.unwrap();

Loading the actual device
It is only possible to load a specific number of items form a specific application. After loading the chirpstack devices to the container, get a device list item out of it and load the device.

 use elorapi::devices::Device;

 // load a limited list of devices in a specific application
 let limit = 10;
 let application_id = 1;
 dev_container.load_chirpstack_device_list(limit, application_id, connection.clone()).await.unwrap();

 // get a device list item from the container
 let device_list_item = dev_container.get_chirpstack_device_list().get(0).unwrap();

 // load the device via its dev_eui
 let device = Device::load_device(&device_list_item.dev_eui, connection.clone()).await.unwrap();

Structs

Device

The representation of a device.

DeviceContainer

To manage Devices.

DeviceProfile

The representation of a device profile.

DeviceProfileContainer

To manage DeviceProfiles and Chirpstack device profiles.

Downlink

To store the structure of downlink messages of a DeviceProfile.

DownlinkPayload

To store the structure of downlink commands of a DeviceProfile .

Uplink

To store the labels of the uplink messages of a DeviceProfile.