Struct elorapi::devices::DeviceProfile

pub struct DeviceProfile {
    pub dev_prof: Option<ChirpstackDeviceProfile>,
    /* private fields */
}

The representation of a device profile.

Fields
dev_prof: Option<ChirpstackDeviceProfile>

Device profile on the Chirpstack Server.

Implementations

impl DeviceProfile

pub fn new(
id: &str,
uplink: Option<Uplink>,
downlink: Option<Downlink>
) -> DeviceProfile

Creates a new device profile.

pub async fn load_device_profile(
device_profile_id: &str,
connection: ChirpstackConnection
) -> Result<Self, Error>

Loads respective Chirpstack device profile via the device profile id and creates a new DeviceProfile without Downlink or Uplink.

pub fn get_downlink(&mut self) -> Option<Downlink>

Gets Downlink.

pub fn read_downlink(&mut self, file: &str) -> Result<(), Error>

Reads Downlink from a json file to an already existing DeviceProfile.

pub fn get_uplink(&mut self) -> Option<Uplink>

Gets Uplink.

pub fn read_uplink(&mut self, file: &str) -> Result<(), Error>

Reads Uplink form a json file to an already existing DeviceProfile.

pub fn read_specification(
file: &str,
network_server_id: i64,
organization_id: i64
) -> Result<DeviceProfile, Error>

Creates DeviceProfile, without dev_prof and client, out of a json file.

pub fn add_client(&mut self, client: Option<DeviceProfileServiceClient<Channel>>)

Adds client.

pub async fn write_device_profile(
&mut self,
connection: ChirpstackConnection
) -> Result<(), Error>

Creates a new device profile in Chirpstack.

pub fn print_downlink(&self)

Prints the Downlink.

pub fn print_uplink(&self)

Prints the Uplink.