Module elorapi::rules

This module is for the management of rules and their execution.

Note that a rule, when started, will always be executed when possible. That means everytime the respective device sends a message and it contains the selected data, the action will be executed. The only exception is when a time condition was set.

RuleGenerator contains functions to start a command line program to generate rules.

In the following the usage of this module will be described with examples.

Examples

Startup

Creating a rule depending on a device

Creating conditions depending on time

Start of a rule

Startup

To use this module it is necessary to create a rule container.

 use std::sync::{Arc, Mutex};
 use elorapi::rules::RuleContainer;

 // create a rule container
 let rule_container_arc_mutex = Arc::new(Mutex::new(RuleContainer::new()));
 let rule_container_arc = Arc::clone(&rule_container_arc_mutex);
 let rule_container = rule_container_arc.lock().unwrap();

Creating a rule depending on a device

There a four steps to follow to create a rule:

Create one or more condition/s
Here you need to get device/s out of the device container (for more information on this see the devices module; here the assumption is made that the device has already been loaded out of the container),
the index of the data you want to compare of the uplink message (this can either be found in the json specification file of the respective device profile, in the uplink message of the device in the chirpstack application sever or by using the function DeviceProfile::print_uplink),
the comparison operator and the value to compare to.

 use elorapi::rules::{Condition, DeviceCondition, RefValue};

 // create first condition
 let index = 2;
 let comparison_operator = "<=".to_string();
 let comparison_value = RefValue::IntNumber(1);
 let condition_one = Condition::Device(DeviceCondition::new(device.clone(), index, comparison_operator, comparison_value));

 // create second condition
 let condition_two = Condition::Device(DeviceCondition::new(device, 4, "==".to_string(), RefValue::String("open".to_string())));

Create an action
To create an action a device is needed, where the action should be executed on,
the indices of the DownlinkPayloads, which should be executed (this information can be found in the specification file or by using the function DeviceProfile::print_downlink),
the message encoded in hexadecimal, which should be sent,
and the port to which the message should be sent to.

 use elorapi::rules::Action;

 let indices = vec![0];
 let message = "message_in_hex".to_string();
 let f_port = 55;
 let action_one = Action::new(device.clone(), indices, message, f_port);

Create the actual rule
Here vectors for all conditions, boolean operators and actions are needed.

 use elorapi::rules::Rule;

 let conditions = vec![condition_one, condition_two];
 let boolean_operators = vec!["&".to_string()];
 let actions = vec![action_one];
 let rule = Rule::new(conditons, boolean_operators, actions);

Add rule to the container

rule_container.add_rule(rule);

Creating conditions depending on time

In this example a rule is created, which should be executed from Monday to Wednesday between 8am and 3pm.

 use chrono::{NaiveTime, Weekday};
 use elorapi::rules::{Condition, TimeCondition};

 // time
 let timespan_start = NaiveTime::from_hms(8, 0, 0);
 let timespan_end = NaiveTime::from_hms(15, 0, 0);

 // condition monday
 let weekday = Some(Weekday::Mon);
 let time_condition = TimeCondition::new(weekday, timespan_start, timespan_end);
 let condition_mon = Condition::Time(time_condition);

 // condition tuesday
 let condition_tue = Condition::Time(TimeCondition::new(Some(Weekday::Tue), timespan_start, timespan_end));

 // condition wednesday
 let condition_tue = Condition::Time(TimeCondition::new(Some(Weekday::Wed), timespan_start, timespan_end));

The next example shows the condition, when you want an action to be executed everyday, but only in a specific timespan. Here between 11pm and 4pm.

 use chrono::NaiveTime;
 use elorapi::rules::{Condition, TimeCondition};

 let timespan_start = NaiveTime::from_hms(23, 0, 0);
 let timespan_end = NaiveTime::from_hms(16, 0, 0);
 let condition = Condition::Time(TimeCondition::new(None, timespan_start, timespan_end));

Start of a rule

It is necessary to establish a Mqtt connection and a ChirpstackConnection, when a rule should be executed. For more information on this see the connections module.

To start all rules you need to have the locked rule container with rules in it,
the established mqtt connection,
the locked device profile container, which contains at least the device profiles used in the rules,
the established container connection and the established Chirpstack connection.

 use elorapi::connections::{ChirpstackConnection, Mqtt};
 use elorapi::rules::RuleContainer;

 // create a Chirpstack connection
 let chirpstack_connection = ChirpstackConnection::new("APIToken", "uri");

 // create a Mqtt connection
 let mqtt = Mqtt::new("uri", "username", "password");
 // this starts receiving the messages from the mqtt broker and publish them into a queue
 let receiver = mqtt.start_receiving();

 // establish container connection
 rule_container.establish_connection(chirpstack_connection.clone()).await.unwrap();

 // get rule container client
 let container_client = rule_container.get_client().unwrap();

 let mut handlers = Vec::new();
 // loop to all rules in the container and start them
 for rule in rule_container.get_rules() {
    // this is necessary due to the loop
    let mqtt_clone = mqtt.clone();
    // this is necessary due to the loop
    let chirpstack_connection_clone = chirpstack_connection.clone();
    // this starts the rule execution of the specific rule
    let handler = RuleContainer::start_rule_execution(rule, mqtt_clone, &device_profile_container, container_client, chirpstack_connection_clone);
    handlers.push(handler);
 }

 // this is necessary that the associated thread is executed
 for join_handle in handlers {
    join_handle.join().unwrap();
 }

Structs

Action

Action which should be executed, when the Conditions of a Rule are true.

DeviceCondition

Condition for a Rule depending on a Device.

Rule

Representation of a rule.

RuleContainer

To manage Rules.

RuleGenerator

To generate a Rule via a command line application.

TimeCondition

Condition for a Rule depending on time.

Enums

Condition

Enum for possible conditions of a Rule.

RefValue

Enum for possible threshold values.